Added rocket sprite and edited angle calculations

+++ src/Projectile.java
@@ -13,8 +13,8 @@ public class Projectile extends Pane {
 	int offsetY = 0;
 	int width;
 	int height;
-	int x; //Proj xPos
-	int y; //Proj yPos
+	double x; //Proj xPos
+	double y; //Proj yPos
 
 	double vx;
 	double vy;
@@ -63,10 +63,10 @@ public class Projectile extends Pane {
 			}
 		}
 
-		this.setTranslateX(this.getTranslateX() + (int) vx);
-		this.setTranslateY(this.getTranslateY() + (int) vy);
-		this.x += (int) vx; //X is int and vx is double 
-		this.y += (int) vy; //casting to int improves missile accuracy
+		this.setTranslateX(this.getTranslateX() + vx);
+		this.setTranslateY(this.getTranslateY() + vy);
+		this.x += vx; //X is int and vx is double 
+		this.y += vy; //casting to int improves missile accuracy
 	}
 
 	public int getVelocityX() {
